fix: Updated URL and add dynamic date in WildBerriesNetworkConnector
The URL in the WildBerriesNetworkConnector class has been updated to include a variable for the current date. Besides, the datetime module has been imported to dynamically generate the current date. The run_validating method was modified to accommodate these changes and pass the URL to the _get_website_response method.