feature: Added SubscriptionDB class to MongoDB package
A new file `subscription.py` was added in the `app/pkg/mongodb/main_cluster` directory. The file contains the `SubscriptionDB` class which interacts with MongoDB to find active subscriptions. The class includes a method `find_active_subscriptions` which constructs a pipeline for aggregation and returns a list of active vendor subscriptions.