feature: Refactored NotificationFabric to include vendor name
The 'create_model' method in the NotificationFabric class is updated to now accept an additional 'vendor_name' parameter. A new static method '_get_message' is added to generate a dynamic message depending on the 'vendor_name'. The 'create_notification' method in the NotificationHandler is also updated to pass 'vendor_name' to 'create_model'.