refactor: Reconfigured database initiation and environment handling
This commit modifies the database initiation process and adjusts how environment variables are handled in the configuration. Previously, both main and local databases were always initiated, but now the local database initiation is conditional based on the "PROD" environment variable. Also, the MongoDB ports no longer need to be explicitly cast to integers.